backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
openai==1.30.1
pdfplumber==0.11.0
docx2txt==0.8
razorpay==1.4.1
python-dotenv==1.0.1
pydantic==2.7.1
